Update(context)=> transfers Cart and Favorites to MainContext
