+++ src/context/userContext.tsx
@@ -2,13 +2,11 @@ import React, { createContext, useContext, useState, useEffect } from 'react';
 
 import { MainContext } from './mainContext';
 
-import { destroyCookie, getCookie, setCookie } from '../helpers/cookie';
+import { getCookie } from '../helpers/cookie';
 import { axiosGetUserInfos } from '../helpers/axios';
-import { decrypt, encrypt } from '../helpers/crypto';
 
 import { PropChild } from '../interfaces/default';
 import { Adress, Card, InfosType, Purchase } from '../interfaces/userInfos';
-import { ProductType } from '../interfaces/store';
 
 export const UserContext = createContext({});
 
@@ -18,8 +16,6 @@ export function UserProvider({ children }: PropChild) {
   const [adresses, setAdresses] = useState([] as Adress[]);
   const [cards, setCards] = useState([] as Card[]);
   const [purchases, setPurchases] = useState([] as Purchase[]);
-  const [cart, setCart] = useState([] as ProductType[]);
-  const [favorites, setFavorites] = useState([] as ProductType[]);
 
   const getUserInfos = async () => {
     const auth = getCookie('auth_handler');
@@ -44,35 +40,6 @@ export function UserProvider({ children }: PropChild) {
     }
   }, [active]);
 
-  useEffect(() => {
-    if (active) {
-      const cartCookie = getCookie('cart_backup');
-      const favoritesCookie = getCookie('user_favorites');
-
-      if (typeof cartCookie === 'string') {
-        const cookieValue = decrypt(cartCookie);
-        setCart(cookieValue);
-      }
-      if (typeof favoritesCookie === 'string') {
-        const cookieValue = decrypt(favoritesCookie);
-        setFavorites(cookieValue);
-      }
-
-    } else {
-      destroyCookie('cart_backup');
-      destroyCookie('user_favorites');
-    }
-
-  }, [active]);
-
-  useEffect(() => {
-    setCookie('cart_backup', encrypt(cart));
-  }, [cart]);
-
-  useEffect(() => {
-    setCookie('user_favorites', encrypt(favorites));
-  }, [favorites]);
-
   const userValue = {
     active,
     setActive,
@@ -80,10 +47,6 @@ export function UserProvider({ children }: PropChild) {
     adresses,
     cards,
     purchases,
-    cart,
-    setCart,
-    favorites,
-    setFavorites,
   };
 
   return (
